Revert handling of negative numbers for now
Adding functionality to the lexer to be able to lex negative numbers has
introduced all sorts of bugs. For example, '1-2' is no longer parsed as
'one minus two' but rather 'the number one followed by the number
negative two', which is obviously wrong. Perhaps negative numbers are
not the responsibility of the lexer?